+++ core/src/main/java/es/urjc/etsii/grafo/exceptions/DefaultExceptionHandler.java
@@ -6,11 +6,13 @@ import es.urjc.etsii.grafo.exception.InvalidRandomException;
 import es.urjc.etsii.grafo.io.Instance;
 import es.urjc.etsii.grafo.services.IOManager;
 import es.urjc.etsii.grafo.solution.Solution;
+import es.urjc.etsii.grafo.util.ExceptionUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Optional;
-import java.util.logging.Logger;
 
 /**
  * Default exception handler.
@@ -21,7 +23,7 @@ import java.util.logging.Logger;
  * @param <I> Instance class
  */
 public class DefaultExceptionHandler<S extends Solution<S,I>, I extends Instance> extends ExceptionHandler<S,I> {
-    private static final Logger logger = Logger.getLogger(DefaultExceptionHandler.class.getName());
+    private static final Logger logger = LoggerFactory.getLogger(DefaultExceptionHandler.class.getName());
 
     private final IOManager<S,I> io;
 
@@ -36,17 +38,18 @@ public class DefaultExceptionHandler<S extends Solution<S,I>, I extends Instance
      * Behaviour can be customized or changed by extending the ExceptionHandler class.
      */
     public void handleException(String experimentName, int iteration, Exception e, Optional<S> sOptional, I i, Algorithm<S,I> algorithm){
-        logger.severe(String.format("Error while solving instance %s with algorithm %s, iteration %s, skipping. Exception message: %s", i.getId(), algorithm.toString(), iteration, e.getMessage()));
+        logger.error("Error while solving instance {} with algorithm {}, iteration {}, skipping. Exception message: {}.", i.getId(), algorithm.toString(), iteration, e.getMessage());
         explain(e);
         String stackTrace = getStackTrace(e);
-        logger.severe("Stacktrace: " + stackTrace);
-        sOptional.ifPresent(s -> logger.severe("Last executed movements: " + s.lastExecutesMoves()));
+        logger.info("Simplified stacktrace: {}", ExceptionUtil.filteredStacktrace(e));
+        logger.trace("Full error stacktrace: {}", stackTrace);
+        sOptional.ifPresent(s -> logger.info("Last executed moves: {}", s.lastExecutesMoves()));
         io.exportError(experimentName, algorithm, i, e, stackTrace);
     }
 
     private void explain(Throwable e) {
         if(e instanceof InvalidRandomException){
-            logger.severe("The exception InvalidRandomException is generated when a method from the Java API that would break reproducibility is executed. See https://mork-optimization.readthedocs.io/en/latest/quickstart/bestpractices/#customized-random-generator for more information.");
+            logger.error("The exception InvalidRandomException is generated when a method from the Java API that would break reproducibility is executed. See https://mork-optimization.readthedocs.io/en/latest/quickstart/bestpractices/#customized-random-generator for more information.");
         }
     }